// src/api/generated/toodlApi.zod.ts
export const listUpdateBodyNameMax = 20;

export const listUpdateBodyColorRegExp = new RegExp('^#[0-9A-Fa-f]{6}$');


export const ListUpdateBody = zod.object({
  "name": zod.string().min(1).max(listUpdateBodyNameMax),
  "color": zod.string().regex(listUpdateBodyColorRegExp),
  "type": zod.enum(['REGULAR', 'SHOPPING']).optional()
})